ingester/pyproject.toml
[project]

name = "ingester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nats-py>=2.9.0",
    "pydantic>=2.9.2",
    "typer>=0.12.5",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "ruff>=0.7.1",
]
